remove all warnings

+++ src/rio.h
@@ -10,7 +10,7 @@
 #define RIO_BUFSIZE 8192
 typedef struct {
     int rio_fd;                /* descriptor for this internal buf */
-    int rio_cnt;               /* unread bytes in internal buf */
+    ssize_t rio_cnt;             /* unread bytes in internal buf */
     char *rio_bufptr;          /* next unread byte in internal buf */
     char rio_buf[RIO_BUFSIZE]; /* internal buffer */
 } rio_t;
